# PERF: get_dummies create array of 0/1s with appropriate layout for internals